requirements.txt
Django
Pillow
Psycopg2-binary
Djangorestframework
django-filter